+++ praw/models/reddit/redditor.py
@@ -144,6 +144,14 @@ class Redditor(MessageableMixin, RedditorListingMixin, FullnameMixin, RedditBase
             self._fullname = fullname
         super().__init__(reddit, _data=_data, _extra_attribute_to_check="_fullname")
 
+    def __setattr__(self, name: str, value: Any):
+        """Objectify the subreddit attribute."""
+        if name == "subreddit" and value:
+            from .user_subreddit import UserSubreddit
+
+            value = UserSubreddit(reddit=self._reddit, _data=value)
+        super().__setattr__(name, value)
+
     def _fetch_username(self, fullname):
         return self._reddit.get(API_PATH["user_by_fullname"], params={"ids": fullname})[
             fullname